feat: show docker image builder pods
