+++ dashboard/package.json
@@ -56,7 +56,7 @@
     "react-dom": "^17.0.2",
     "react-hooks-global-state": "^1.0.1",
     "react-i18next": "^11.11.4",
-    "react-icons": "^4.2.0",
+    "react-icons": "^4.4.0",
     "react-jss": "^10.7.1",
     "react-lazylog": "^4.5.3",
     "react-query": "^3.19.6",
